Refactor: TweetHelperをSnsHelperに変更した
Twitterに限らずfacebookの内容もいれるつもりなので、より汎用的な名前にした。
